Issue #116 and revisiting #112
issue 116 stemmed from a misunderstanding of subprocess protocols in
asyncio. Even though I fixed the test I decided to ultimatlyu remove it
since it was a test of my own understanding of asyncio and not a valid
unit tests of Nanaimo.

After understanding the subprocess protocols better I realized that #
112 wasn't fixed properly. I have revisited this issue and provided a
superior fix.